Add missing dependencies and untangle interdependencies
1) Adds missing #includes for headers. #include graphs of source files
remain unchanged (including csalerts.mm).

2) Untangles scottish_textures.cpp and low_level_textures.h from each
other by moving some code from the former into the latter. In the
process, a) the completely redundant second set of definitions for the
HORIZONTAL_* macros is removed, b) the texture_random_seed global is
converted to a reference-returning inline function so that its
definition and usage can be contained entirely within
low_level_textures.h, and c) the "#include low_level_textures.h" line is
moved up to join its #include line buddies.

3) Untangles render.h and scottish_textures.h from each other by moving
some code from the former into the latter. In the process, the redundant
point2d typedef is removed and the "#include scottish_textures.h" line
is moved up to join its #include line buddies.
